refactor: update property names in ImpactInteraction and ImpactMetric to camelCase
