+++ src/api/models/ImpactMetric.ts
@@ -2,7 +2,7 @@
 /* istanbul ignore file */
 /* tslint:disable */
 /* eslint-disable */
-import type { ImpactInteraction } from 'src/api';
+import type { ImpactInteraction } from './ImpactInteraction';
 export type ImpactMetric = {
   applicationName: string;
   featureName: string;
@@ -11,7 +11,7 @@ export type ImpactMetric = {
   startTimeStamp?: string;
   endTimeStamp?: string;
   clickCount: number;
-  Interactions?: Array<ImpactInteraction> | null;
+  interactions?: Array<ImpactInteraction> | null;
   userId: string;
   isSuccessful?: boolean | null;
 };
